scripts/requirements.txt
cassandra-driver==3.29.2
click==8.2.1
Faker==37.4.0
geomet==0.2.1.post1
neo4j==5.28.1
numpy==2.3.1
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
